+++ Frends.Mllp.Receive/Frends.Mllp.Receive/Definitions/Connection.cs
@@ -75,7 +75,6 @@ public class Connection
     /// </summary>
     /// <example>E5FA62B8B5F3B0B2B3B4B5B6B7B8B9B0B1B2B3B</example>
     [DisplayFormat(DataFormatString = "Text")]
-    [DefaultValue("")]
     [UIHint(nameof(TlsMode), "", TlsMode.Mtls)]
     public string[] ClientCertificateThumbprints { get; set; } = [];