+++ Frends.Mllp.Receive/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 
-- 'New' `Connection.ClientCertificateThumbprints` property to specify client certificate thumbprints for authentication in MTLS mode
-- 
+- `Connection.ClientCertificateThumbprints` property to specify client certificate thumbprints for authentication in
+  MTLS mode
+
 ## [1.1.0] - 2026-04-13
 
 ### Added